requirements.txt
discord.py==1.6.0
webrepl==0.2.0